+++ team08/index.js
@@ -1,4 +1,4 @@
-import { loaddb } from "../scripts/vocabulary_await.js";
+import { loaddb } from "./store/alternative_backend/vocabulary_await.js";
 // ==============================================
 // Owned by Team 08
 // ==============================================
